Damage system now works

+++ include/collision_component.h
@@ -19,7 +19,7 @@ class CollisionComponent: public IComponent {
         return body;
     }
 
-    void start_collision(size_t collide_id) {
+    void start_collision(uint16_t collide_id) {
         is_collision = true;
         this->collide_id = collide_id;
         update();
@@ -33,14 +33,14 @@ class CollisionComponent: public IComponent {
         return is_collision;
     }
 
-    size_t get_collide_id() {
+    uint16_t get_collide_id() {
         return collide_id;
     }
 
   private:
     b2Body* body;
     bool is_collision = false;
-    size_t collide_id = 0;
+    uint16_t collide_id = 0;
 };
 
 #endif  // ACID_INCLUDE_COLLISION_COMPONENT_H_
